+++ scripts/macros.js
@@ -159,6 +159,7 @@ export {elementalAdeptA, elementalAdeptC, elementalAdeptF, elementalAdeptL, elem
 export {grappler} from './macros/2024/feats/grappler.js';
 export {heavyArmorMaster} from './macros/2024/feats/heavyArmorMaster.js';
 export {piercer} from './macros/2024/feats/piercer.js';
+export {savageAttacker} from './macros/2024/feats/savageAttacker.js';
 //Class Features
 //Barbarian
 export {rage, rageRaging, rageUpkeep} from './macros/2024/classFeatures/barbarian/rage.js';